tests/wpt/tests/css/css-scroll-anchoring/inheritance.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>Inheritance of CSS Scroll Anchoring properties</title>
<link rel="help" href="https://drafts.csswg.org/css-scroll-anchoring/#property-index">
<meta name="assert" content="overflow-anchor does not inherit.">
<meta name="assert" content="overflow-anchor has initial value 'none'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/inheritance-testcommon.js"></script>
</head>
<body>
<div id="container">
  <div id="target"></div>
</div>
<script>
assert_not_inherited('overflow-anchor', 'auto', 'none');
</script>
</body>
</html>
